dev: tests refactoring
